// Frontend/src/app/modules/ct-client/model/product-search-request.ts
/**
 * ProductManagementAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductSearchRequest { 
    pageNumber?: number;
    pageSize?: number;
    sort_by?: string | null;
    descending?: boolean;
    name?: string | null;
    categoryId?: string | null;
    lastUpdatedStart?: Date | null;
    lastUpdatedEnd?: Date | null;
    sku?: string | null;
    inStock?: boolean | null;
}